% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An Example Trial Dataset}
\format{
A data frame with 3 variables and 200 observations
}
\usage{
data(example_data)
}
\description{
A dataset containing the time-to-event information. The variables are as follows
}
\details{
\itemize{
 \item time. event or censoring time
 \item censor. censoring indicator; 1=event
 \item Trt. treatment arm name

}
}
\keyword{datasets}
